Improved Support for Textless ROM
Added the MD5 Hash of the textless ROM to randomizer.py, allowing the program to recognize the textless ROM as valid.

The program should be able to identify the textless ROM as a valid FF6 ROM file when scanning the folder.

More importantly, the textless ROM no longer generates the warning message about not matching known MD5 Hashes, requiring the user to answer a prompt to continue.